Data Cleaning
I took all of the data and grouped it by each day and then took only the row containing the max temperature value for that day. I then noticed that in the original data that there are many duplicate months thrown in there all over the place so I cleaned those up. The daily_high_temps_distinct is the file that has the final product while the daily_high_temps still has all of the duplicate rows.